// src/types/anime.ts
export type Anime = {
  title?: string | undefined;
  img?: string | undefined;
  sinopsis?: string | undefined;
  info?: {
    Tipo: string | undefined;
    Genero: string | undefined;
    Studios: string | undefined;
    Demografia: string | undefined;
    Idiomas: string | undefined;
    Episodios: string | undefined;
    Duracion: string | undefined;
    Emitido: string | undefined;
    Estado: string | undefined;
    Calidad: string | undefined;
  };
};